Injected Backend logic into PrepareGift2.html
Redirected PrepareGift1.html to PrepareGift2.html

Directed PrepareGift2.html to PrepareGift3.html

Modified PrepareGift2.html and PrepareGift2.js to run together

Changed PrepareGift2.js to update players instead of overwriting them

Implemented getPlayerName function to retrieve a player's name from the database, given the playerID. The playerID is obtained by parsing the URL.

NOTE: Players are currently able to move to PrepareGift3.html without choosing a box and ribbon color. This is because of a bug that occurs when trying to prevent the user from proceeding. When the database is referenced in the event handler, any code written inside of the database callback function repetitively runs.