Fixed countdown timer bug in order.js
The countdown timer function was not working properly. The countdown tick interval was not static as it was dependent on a dynamic variable. In addition, the countdown function was being called multiple times, depending on the number of players in the game, which also accelerated the countdown. These issues have been addressed and the count down timer is working properly now.

I have also cleaned up the code and comments